update index.html
In your `header` section, you have a `div class = logo` that wraps the profile picture. However, the class is not declared in your style.css file. I would suggest that you create a CSS rule for that class, and add a little margin of like `5 or 10px` to it so that the profile picture does not completely reach the top.